[fix] random 2202 error → send the hash salt 'nonce' as string
This makes some workarounds obsolete. The API documentation file xmlrpc.pdf
stated that the nonce value had to be a float value. This leads, however, to
problems, as the nonce variable must be interpreted as string to calculate
the hash. This has to be done by the client and by the server and that leads
to failures when the have different string representation for the float
valued nonce.  Using a string value as nonce is not what the API
documentation says but it is the way, the Java example file is built. Thus,
the use of a string as nonce should be fine.

The workarounds commits made obsolete now:
  * a0e796cb77e6c70836ff Random errors occuring...
  * acde9e2eacf2d1a4d18a if a random 2202 error occurs, rehash the password...